Blockchain_AntiCounterFeit_ver1/requirement.txt
eth-brownie
qrcode

